Backend/Backend/requirements.txt
sentence-transformers==5.1.2
faiss-cpu==1.12.0
datasets==4.3.0
transformers==4.57.1
torch==2.5.1
numpy>=1.26.0,<2.0
scikit-learn==1.7.2
scipy==1.16.2
huggingface-hub==0.36.0
tokenizers==0.22.1
safetensors==0.6.2
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
sacremoses==0.1.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
rank-bm25==0.2.2
nltk==3.9.1
